requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
qdrant-client==1.11.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
